feat: delete points from a line
it however has a bug. Both adding a new point and removing one work on double click so every time you remove a point a new one is added to that position. The problem is not because of the way deleting points works so I just have to implement a better way of adding new points.